started making chests
the work of chests are in progress
the texture is getting displayed and the blocks is placeable. for it to work are a progress. maybe i need to change the hole block code to make it work with pointers? or i need another approach. just an important thing is i need to figure out what method is best for storing data without to much lag.